Make hamming and difference reifiable, add constraints to tester. Minor changes to gcc and gacalldiff
git-svn-id: svn+ssh://keith.cs.st-and.ac.uk/mnt/raid/svn/repositories/minion-private/public/trunk@2205 477b24dd-e91e-0410-aa49-980f3df0729b

+++ minion/constraints/gcc_common.h
@@ -153,7 +153,7 @@ struct GCC : public AbstractConstraint
                 adjlistpos[i].resize(numvars);
                 for(int j=0; j<numvars; j++) adjlistpos[i][j]=j;
             }
-            adjlistlength=getMemory(stateObj).backTrack().template requestArray<short>(numvars+numvals);
+            adjlistlength=getMemory(stateObj).backTrack().template requestArray<int>(numvars+numvals);
             for(int i=0; i<numvars; i++) adjlistlength[i]=numvals;
             for(int i=numvars; i<numvars+numvals; i++) adjlistlength[i]=numvars;
         #endif
@@ -172,7 +172,7 @@ struct GCC : public AbstractConstraint
     
     #ifdef INCGRAPH
     vector<vector<int> > adjlist;
-    MoveableArray<short> adjlistlength;
+    MoveableArray<int> adjlistlength;
     vector<vector<int> > adjlistpos;   // position of a variable in adjlist.
     #endif
